fix: factor out concurrency queries into overwrite
